+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # GG Changelog
 
+## Unreleased
+
+### Added
+- Added git remotes to the status bar, with push & fetch commands.
+
 ## [0.15.2](releases/tag/v0.15.2)
 
 ### Fixed